Ajustando cronograma e alterando tipos da Atividade 3, de int para double

+++ 2018.1/atividades/03-SimpleInterpreter/ast-simple-interpreter/src/br/ufpe/cin/if688/ast/NumExp.java
@@ -3,9 +3,9 @@ package br.ufpe.cin.if688.ast;
 import br.ufpe.cin.if688.visitor.IVisitor;
 
 public class NumExp extends Exp {
-	int num;
+	double num;
 
-	public NumExp(int n) {
+	public NumExp(double n) {
 		num = n;
 	}
